requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.24.0
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
python-dotenv==1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
